requirements.txt
fastapi
uvicorn[standard]
azure-search-documents
openai
python-dotenv
langchain_openai
langchain_core
langchain-community
